Initial build of Java application.
1) Creates a main window. 
2) Scans for network devices (Default device is set in the NetworkingInfo class. If your ip starts with something other than 192.168.1, change this here. Needs to be *more* dynamic in the future. I will explain on Tuesday.)
3) Based on the network device, scans that subnet for devices (if your ip is 192.168.1.68, it will scan 192.168.1.0 through 192.168.1.255). 
4) Puts all found network devices in a list, stored in NetworkingInfo.java
5) Populates a JList with network devices found
6) Adds functionality of the JList that prints the device name. 

Any devices found on the network are saved. From here we can export to database (up to you, I'm familiar with access but I know it sucks). We can then use PHP / Javascript to pull from the database to display in d3. 

Recommend taking  a few minutes on Tuesday (10/22/2013) to go over initial design and then our final design parameters.