Added menu item to select IPv4 or IPv6. ! This is not implemented yet, it basically does nothing !
This is a placeholder. Set Port scanning back up. Application crashes due to what is *probably* a cross-threading error in ScanNetworkForDevices in the Port scanning error. I think it's happening when I try to add the port to an array. 

Will try to at least get this part fixed and then an example output file produced by tomorrow or Tuesday at the latest. 

If you want to work and the application keeps crashing, just comment out the part where I fire up the port scanner.