added parse_data folder and contents
I included the cve zip, so that adds 14mb. Feel free to remove

parse_cve.pl takes the cve file and cuts it down to a specified range
of years and by a specified set of services.

It also outputs a clean.csv that rates the services in terms of risk.